nested metadata_extensions: https://github.com/metanorma/metanorma-bipm/issues/1
