+++ lib/asciidoctor/generic/converter.rb
@@ -1,13 +1,10 @@
 require "asciidoctor"
 require "asciidoctor/standoc/converter"
 require "fileutils"
+require_relative "front"
 
 module Asciidoctor
   module Generic
-
-    # A {Converter} implementation that generates RSD output, and a document
-    # schema encapsulation of the document for validation
-    #
     class Converter < Standoc::Converter
       XML_ROOT_TAG = "generic-standard".freeze
       XML_NAMESPACE = "https://www.metanorma.org/ns/generic".freeze
@@ -28,110 +25,12 @@ module Asciidoctor
           self.class::_file || __FILE__), "..", "..", "..", loc))
       end
 
-      def default_publisher
-        configuration.organization_name_long
-      end
-
-      def org_abbrev
-        if !configuration.organization_name_long.empty? &&
-            !configuration.organization_name_short.empty? &&
-            configuration.organization_name_long !=
-            configuration.organization_name_short
-          { configuration.organization_name_long =>
-            configuration.organization_name_short }
-        else
-          super
-        end
-      end
-
-      def relaton_relations
-        Array(configuration.relations) || []
-      end
-
-      def metadata_committee(node, xml)
-        return unless node.attr("committee")
-        xml.editorialgroup do |a|
-          a.committee node.attr("committee"),
-            **attr_code(type: node.attr("committee-type"))
-          i = 2
-          while node.attr("committee_#{i}") do
-            a.committee node.attr("committee_#{i}"),
-              **attr_code(type: node.attr("committee-type_#{i}"))
-            i += 1
-          end
-        end
-      end
-
-      def metadata_status(node, xml)
-        xml.status do |s|
-          s.stage ( node.attr("status") || node.attr("docstage") ||
-                   configuration.default_stage || "published" )
-          x = node.attr("substage") and s.substage x
-          x = node.attr("iteration") and s.iteration x
-        end
-      end
-
       def docidentifier_cleanup(xmldoc)
         template = configuration.docid_template ||
           "{{ organization_name_short }} {{ docnumeric }}"
         docid = xmldoc.at("//bibdata/docidentifier")
         id = boilerplate_isodoc(xmldoc).populate_template(template, nil)
         id.empty? and docid.remove or docid.children = id
-      end
-
-      def metadata_id(node, xml)
-        xml.docidentifier **{ type:
-                              configuration.organization_name_short } do |i|
-          i << "DUMMY"
-        end
-        xml.docnumber { |i| i << node.attr("docnumber") }
-      end
-
-      def metadata_ext(node, ext)
-        super
-        if configuration.metadata_extensions.is_a? Hash
-          metadata_ext_hash(node, ext, configuration.metadata_extensions)
-        else
-          Array(configuration.metadata_extensions).each do |e|
-            a = node.attr(e) and ext.send e, a
-          end
-        end
-      end
-
-      EXT_STRUCT = %w(_output _attribute _list).freeze
-
-      def metadata_ext_hash(node, ext, hash)
-        hash.each do |k, v|
-          next if EXT_STRUCT.include?(k) || !v&.is_a?(Hash) && !node.attr(k)
-          if v&.is_a?(Hash) && v["_list"]
-            csv_split(node.attr(k)).each do |val|
-              metadata_ext_hash1(k, val, ext, v, node)
-            end
-          else
-            metadata_ext_hash1(k, node.attr(k), ext, v, node)
-          end
-        end
-      end
-
-      def metadata_ext_hash1(key, value, ext, hash, node)
-        return if hash&.is_a?(Hash) && hash["_attribute"]
-        is_hash = hash&.is_a?(Hash) &&
-            !hash.keys.reject { |n| EXT_STRUCT.include?(n) }.empty?
-        return if !is_hash && (value.nil? || value.empty?)
-        name = hash&.is_a?(Hash) ? (hash["_output"] || key) : key
-        ext.send name, **attr_code(metadata_ext_attrs(hash, node)) do |e|
-            is_hash ? metadata_ext_hash(node, e, hash) : (e << value)
-        end
-      end
-
-      def metadata_ext_attrs(hash, node)
-        return {} unless hash.is_a?(Hash)
-        ret = {}
-        hash.each do |k, v|
-          next unless v.is_a?(Hash) && v["_attribute"]
-          ret[(v["_output"] || k).to_sym] = node.attr(k)
-        end
-        ret
       end
 
       def doctype(node)
@@ -260,6 +159,18 @@ module Asciidoctor
         f.is_a? String and return baselocation(f)
         f.is_a? Hash and f[@lang] and return baselocation(f[@lang])
         super
+      end
+
+      def cleanup(xmldoc)
+        super
+        empty_metadata_cleanup(xmldoc)
+        xmldoc
+      end
+
+      def empty_metadata_cleanup(xmldoc)
+        xmldoc.xpath("//bibdata/ext//*").each do |x|
+          x.remove if x.children.empty?
+        end
       end
     end
   end
